reorganized client file structure relates #36
