--- client/src/index.css
@@ -1,31 +1,0 @@
-body {
-  margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen",
-    "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue",
-    sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
-    monospace;
-}
-
-.chat-container {
-  width: 80vw;
-  height: 50vh;
-  padding: 10px;
-  border: 1px solid white;
-  display: flex;
-  flex-direction: column;
-  overflow-y: scroll;
-}
-
-.chat {
-  margin: 5px;
-  padding: 5px;
-  font-size: 3.2vw;
-  color: white;
-  border: 1px solid white;
-}
