+++ client/src/index.js
@@ -3,7 +3,6 @@ import ReactDOM from 'react-dom';
 import { BrowserRouter as Router } from 'react-router-dom';
 
 import App from './App';
-import './index.css';
 
 ReactDOM.render(
   <Router>
