+++ client/src/pages/ChattingRoom/ChattingRoom.css
@@ -1,0 +1,17 @@
+.chat-container {
+  width: 80vw;
+  height: 50vh;
+  padding: 10px;
+  border: 1px solid white;
+  display: flex;
+  flex-direction: column;
+  overflow-y: scroll;
+}
+
+.chat {
+  margin: 5px;
+  padding: 5px;
+  font-size: 3.2vw;
+  color: white;
+  border: 1px solid white;
+}
