+++ client/src/pages/index.js
@@ -1,0 +1,5 @@
+export { default as Home } from './Home';
+export { default as Signup } from './Signup';
+export { default as Login } from './Login';
+export { default as Rooms } from './Rooms';
+export { default as ChattingRoom } from './ChattingRoom';
